application-code/container-queue-proc/src/requirements.txt
boto3==1.24.58
Pillow==9.2.0
